requirements.txt
fastapi
uvicorn
python-multipart
aioboto3
pydantic-settings
sqlalchemy
aiosqlite